usb: modify usb_gadget_handle_interrupts to take controller index
Since we support multiple dwc3 controllers to be existent at the same
time, in order to handle the interrupts of a particular dwc3 controller
usb_gadget_handle_interrutps should take controller index as an
argument.

Hence the API of usb_gadget_handle_interrupts is modified to take
controller index as an argument and made the corresponding changes to all
the usb_gadget_handle_interrupts calls.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Reviewed-by: Lukasz Majewski <l.majewski@samsung.com>

Conflicts:
	board/ti/am43xx/board.c
	board/ti/dra7xx/evm.c
	common/cmd_fastboot.c
	drivers/usb/gadget/atmel_usba_udc.c
	drivers/usb/gadget/ci_udc.c
	drivers/usb/gadget/fotg210.c