common: decompress_ext4: support multiple chain downloads
This patch supports multiple chain downloads which splits a large buffer to
small buffers. The sparse image of android consists of file header,
chunk header and raw data. The file header is the first 28 bytes of the
buffer and it indicates meta information like total size, total chunk
count. Following the file header, there is multiple chunks which has a
chunk header and raw data. The types of the chunks are RAW, FILL and
NONE. The RAW type has raw data in the chunk and the FILL type means
there is no raw data thus we could skip the sectors.

The original scheme of fastboot is flashing after we download the whole
data. However, there is no way to store a huge image if we don't have
sufficient ram buffers. Thus, we need to flash the image into small size
buffers. To support chain download, we need to keep meta information
like current sector position.

|------------------------80MB---------------------|
|-----32MB---------|-------32MB-----|----16MB-----|
|FH|CH|----RAW------|CH|FILL|CH|-------RAW--------|

For example, there is 80MB image to download and flash, we could split
them into three chunks which the max is 32MB or predefined size. We
should consider residue bytes in boundary of chunks. The residue bytes
should be stored a buffer and flash them when next chunks will be
flashed.

Change-Id: Idabb738fa1c2fac1b1e2c561e3cbdb7fa5d5ac8f
Signed-off-by: Chanho Park <chanho61.park@samsung.com>