common: fastboot: support chunk download
The original policy of the fastboot is first download and after flash.
We can't flash a huge size image because some boards don't have enough
size of ram buffer. Thus, we need to split the image into small size
buffer. Unfortunately, we can't get the partition information in normal
way. But, we could get the info through partition-type command of
fastboot. The partition information should be stored during
partition-type command and use them when download stage. If download
bytes reach chunk size for example 64MB, we could flash the buffer into
storage.

|--------------------192MB------------------|
|-----64MB-----|-----64MB-----|-----64MB----|
|=======A======|=======B======|======C======|

If 'A' chunk is downloaded, the 'A' will be flashed into mmc partition.
After B and C will be downloaded and flashed, the whole download will be
done.

Change-Id: If1321b8660103b7a18fa60ae6631b09e8e171464
Signed-off-by: Chanho Park <chanho61.park@samsung.com>