artik: common: enable factory_info command
This patch enables the factory_info command which is to
retrieve some factory information such as board S/N.
The default location to store the data is early offset of eMMC
because we don't use the location for any other purpose.

"run factory_load" command will load the information from eMMC
and configure them into environment variables.
"ethaddr" is ethernet mac address and it will be passed to kernel.
"bd_addr" is bluetooth address and it also be passed to kernel.

Change-Id: Ifa5660c3e765be84d00732593ca2005a5fc9f749
Signed-off-by: Chanho Park <chanho61.park@samsung.com>